move iterator to tools/roiextractors
